+++ src/features/game/single/main/views/PhotoMainView.vue
