+++ src/router/adminRoutes.js
@@ -1,10 +1,10 @@
-import Dashboard from '@/features/admin/views/DashboardView.vue'
+import DashboardView from '@/features/admin/views/DashboardView.vue'
 
 const adminRoutes = [
     {
         path: '/admin',
         name: "Dashboard",
-        component: Dashboard
+        component: DashboardView
     },
 ];
